refactor: extract formatValue to utility module - Extract formatValue to web/src/util/format.ts to allow sharing and resolve react-refresh/only-export-components linting error. - Update PublicPieceShell.tsx and ShowcaseViewFormat.test.tsx to import from the new utility. Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
