+++ web/src/components/PublicPieceShell.tsx
@@ -1,5 +1,6 @@
 import { useParams } from "react-router-dom";
 import { Box, CircularProgress, Container, Typography, alpha, Divider } from "@mui/material";
+import { formatValue } from "../util/format";
 import { fetchPiece } from "../util/api";
 import { useAsync } from "../util/useAsync";
 import { type PieceDetail } from "../util/types";
@@ -139,19 +140,4 @@ function ShowcaseView({ piece }: { piece: PieceDetail }) {
   );
 }
 
-export function formatValue(value: unknown): string {
-  if (value === null || value === undefined || value === "") {
-    return "";
-  }
-  if (typeof value === "string" || typeof value === "number") {
-    return String(value);
-  }
-  if (typeof value === "boolean") {
-    return value ? "Yes" : "No";
-  }
-  if (typeof value === "object" && value !== null && "name" in value) {
-    const name = (value as { name?: unknown }).name;
-    return typeof name === "string" ? name : "";
-  }
-  throw new Error("Unsupported value type");
-}
+
